Added config directory and moved express setup via app.ts into the config directory.
